requirements.txt
amqp==5.1.0
asgiref==3.5.0
Babel==2.9.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
chardet==4.0.0
Django==3.2.5
django-debug-toolbar==3.2.4
django-maintenance-mode==0.16.2
django-webpack-loader==1.4.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
kombu==5.2.4
Pillow==9.0.1
psycopg2==2.9.3
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tornado==6.1
vine==5.0.0
zipp==3.7.0
matplotlib==3.6.2
numpy==1.21.2
seaborn==0.11.2
scipy==1.10.0
urllib== 3.11.1